+++ apps/prairielearn/src/pages/instructorInstanceAdminSettings/instructorInstanceAdminSettings.html.tsx
@@ -1,11 +1,10 @@
 import clsx from 'clsx';
-import { useState } from 'preact/compat';
-import { Button, Form, InputGroup, OverlayTrigger, Tooltip } from 'react-bootstrap';
+import { Form } from 'react-bootstrap';
 import { useForm } from 'react-hook-form';
 
 import { GitHubButton } from '../../components/GitHubButton.js';
+import { PublicLinkSharing, StudentLinkSharing } from '../../components/LinkSharing.js';
 import type { NavPage } from '../../components/Navbar.types.js';
-import { QRCodeModal } from '../../components/QRCodeModal.js';
 import type { StaffCourseInstanceContext } from '../../lib/client/page-context.js';
 import { type Timezone, formatTimezone } from '../../lib/timezone.shared.js';
 import { encodePathNoNormalize } from '../../lib/uri-util.shared.js';
@@ -18,120 +17,6 @@ interface SettingsFormValues {
   display_timezone: string;
   group_assessments_by: StaffCourseInstanceContext['course_instance']['assessments_group_by'];
   hide_in_enroll_page: boolean;
-}
-
-async function copyToClipboard(text: string) {
-  await navigator.clipboard.writeText(text);
-}
-
-function StudentLinkSharing({
-  studentLink,
-  studentLinkMessage,
-}: {
-  studentLink: string;
-  studentLinkMessage: string;
-}) {
-  const [showQR, setShowQR] = useState(false);
-  const [copied, setCopied] = useState(false);
-  return (
-    <div class="mb-3">
-      <label class="form-label" for="student_link">
-        Student Link
-      </label>
-      <InputGroup>
-        <Form.Control type="text" id="student_link" value={studentLink} disabled />
-        <OverlayTrigger overlay={<Tooltip>{copied ? 'Copied!' : 'Copy'}</Tooltip>}>
-          <Button
-            size="sm"
-            variant="outline-secondary"
-            aria-label="Copy student link"
-            onClick={async () => {
-              await copyToClipboard(studentLink);
-              setCopied(true);
-              setTimeout(() => setCopied(false), 1500);
-            }}
-          >
-            <i class="bi bi-clipboard" />
-          </Button>
-        </OverlayTrigger>
-        <OverlayTrigger overlay={<Tooltip>View QR Code</Tooltip>}>
-          <Button
-            size="sm"
-            variant="outline-secondary"
-            aria-label="Student Link QR Code"
-            onClick={() => setShowQR(true)}
-          >
-            <i class="bi bi-qr-code-scan" />
-          </Button>
-        </OverlayTrigger>
-      </InputGroup>
-      <small class="form-text text-muted">{studentLinkMessage}</small>
-      <QRCodeModal
-        id="studentLinkModal"
-        title="Student Link QR Code"
-        content={studentLink}
-        show={showQR}
-        onHide={() => setShowQR(false)}
-      />
-    </div>
-  );
-}
-
-function PublicLinkSharing({
-  publicLink,
-  sharingMessage,
-  publicLinkMessage,
-}: {
-  publicLink: string;
-  sharingMessage: string;
-  publicLinkMessage: string;
-}) {
-  const [showQR, setShowQR] = useState(false);
-  const [copied, setCopied] = useState(false);
-  return (
-    <>
-      <p>
-        <span class="badge color-green3 me-1">Public source</span>
-        {sharingMessage}
-      </p>
-      <div class="mb-3">
-        <label for="publicLink">Public link</label>
-        <InputGroup>
-          <Form.Control type="text" id="publicLink" value={publicLink} disabled />
-          <OverlayTrigger overlay={<Tooltip>{copied ? 'Copied!' : 'Copy'}</Tooltip>}>
-            <Button
-              size="sm"
-              variant="outline-secondary"
-              aria-label="Copy public link"
-              onClick={async () => {
-                await copyToClipboard(publicLink);
-                setCopied(true);
-                setTimeout(() => setCopied(false), 1500);
-              }}
-            >
-              <i class="far fa-clipboard" />
-            </Button>
-          </OverlayTrigger>
-          <Button
-            size="sm"
-            variant="outline-secondary"
-            aria-label="Public Link QR Code"
-            onClick={() => setShowQR(true)}
-          >
-            <i class="fas fa-qrcode" />
-          </Button>
-        </InputGroup>
-        <small class="form-text text-muted">{publicLinkMessage}</small>
-      </div>
-      <QRCodeModal
-        id="publicLinkModal"
-        title="Public Link QR Code"
-        content={publicLink}
-        show={showQR}
-        onHide={() => setShowQR(false)}
-      />
-    </>
-  );
 }
 
 export function InstructorInstanceAdminSettings({
